PENV-251: split error, bloat, functests
