+++ application/functest/main_func_test.go
@@ -3,7 +3,7 @@
 // This material is licensed under the Insolar License version 1.0,
 // available at https://github.com/insolar/insolar/blob/master/LICENSE.md.
 
-// +build functest bloattest
+// +build functest bloattest functest_error
 
 package functest
